Fix some more camel- to snake-case options
Some of the flake8 options are expressed with camel case in
python-lsp-server configuration. This causes the plugin to pass the
flake8 executable some inexisting options, thus resulting in errors.
With this commit the configuration options are converted to snake case
with a custom hardcoded mapping before constructing the argument string.